pyproject.toml
[project]
name = "stable-diffusion-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
